+++ wgpu-core/src/command/compute.rs
@@ -354,6 +354,13 @@ impl Global {
                 }
             }
 
+            if beginning_of_pass_write_index
+                .or(end_of_pass_write_index)
+                .is_none()
+            {
+                return make_err(CommandEncoderError::TimestampWriteIndicesMissing, arc_desc);
+            }
+
             Some(ArcPassTimestampWrites {
                 query_set,
                 beginning_of_pass_write_index,